CODE: don't use long. possible 64bit multiplayer grief
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@7527 8aca7d54-2c30-db11-9de9-000461428c89

+++ gui/components/gui_textinput.h
@@ -81,7 +81,7 @@ protected:
 	 * reference time for regulating cursor blinking
 	 * @author Knightly
 	 */
-	unsigned long cursor_reference_time;
+	uint32 cursor_reference_time;
 
 	/**
 	 * whether focus has been received
